gcloud/vision/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.0.3
functions-framework==3.0.0
google-api-core==2.7.1
google-auth==2.6.0
google-cloud==0.34.0
google-cloud-vision==2.7.1
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
watchdog==1.0.2
Werkzeug==2.0.3
